Displaying the sizes of types in c

+++ 0x00-hello_world/6-size.c
@@ -7,16 +7,15 @@
  *
  * The program aims to display the size of differnt types 
  *
- * The command used was printf and sizeof
  */
 
 int main(void)
 {
-	printf("Size of a char: %lu byte(s)\n", sizeof(char));
-		printf("size of an int: %lu byte(s)\n", sizeof(int));
-		printf("size of a long int: %lu byte(s)\n", sizeof(long int));
-		printf("size of a long long int: %lu byte(s)\n", sizeof(long long));
-		printf("size of a float: %lu byte(s)\n", sizeof(float));
-		return(0);
+	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(char));
+	printf("size of an int: %lu byte(s)\n", (unsigned long)sizeof(int));
+	printf("size of a long int: %lu byte(s)\n", (unsigned long)sizeof(long int));
+	printf("size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(long long));
+	printf("size of a float: %lu byte(s)\n", (unsigned long)sizeof(float));
+	return(0);
 }
